Move Contants to shared
