+++ plugins/plugin-machine/che-plugin-machine-ext-client/pom.xml
@@ -155,7 +155,7 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
-            <artifactId>che-core-api-project</artifactId>
+            <artifactId>che-core-api-project-shared</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
