--- wsagent/che-core-api-project-share/src/main/java/org/eclipse/che/api/project/shared/Constants.java
@@ -1,23 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2016 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.api.project.shared;
-
-/**
- * @author Vitalii Parfonov
- */
-public class Constants {
-
-    public static final String BLANK_ID                  = "blank";
-    public static final String VCS_PROVIDER_NAME         = "vcs.provider.name";
-
-    private Constants() {
-    }
-}
