+++ wsagent/che-core-api-project-shared/src/main/java/org/eclipse/che/api/project/shared/Constants.java
@@ -15,8 +15,23 @@ package org.eclipse.che.api.project.shared;
  */
 public class Constants {
 
-    public static final String BLANK_ID                  = "blank";
-    public static final String VCS_PROVIDER_NAME         = "vcs.provider.name";
+    public static final String BLANK_ID                = "blank";
+    public static final String ZIP_IMPORTER_ID         = "zip";
+    public static final String VCS_PROVIDER_NAME       = "vcs.provider.name";
+    // rels for known project links
+    public static final String LINK_REL_GET_PROJECTS   = "get projects";
+    public static final String LINK_REL_CREATE_PROJECT = "create project";
+    public static final String LINK_REL_UPDATE_PROJECT = "update project";
+    public static final String LINK_REL_EXPORT_ZIP     = "zipball sources";
+    public static final String LINK_REL_CHILDREN       = "children";
+    public static final String LINK_REL_TREE           = "tree";
+    public static final String LINK_REL_DELETE         = "delete";
+    public static final String LINK_REL_GET_CONTENT    = "get content";
+    public static final String LINK_REL_UPDATE_CONTENT = "update content";
+
+    public static final String LINK_REL_PROJECT_TYPES = "project types";
+
+    public static final String CODENVY_DIR = ".codenvy";
 
     private Constants() {
     }
