+++ wsagent/che-core-api-project/src/main/java/org/eclipse/che/api/vfs/impl/file/DefaultFileWatcherNotificationHandler.java
@@ -11,6 +11,7 @@
 package org.eclipse.che.api.vfs.impl.file;
 
 import org.eclipse.che.api.core.ServerException;
+import org.eclipse.che.api.project.shared.dto.event.FileWatcherEventType;
 import org.eclipse.che.api.vfs.Path;
 import org.eclipse.che.api.vfs.VirtualFile;
 import org.eclipse.che.api.vfs.VirtualFileSystemProvider;
