+++ wsagent/che-core-api-project/src/main/java/org/eclipse/che/api/vfs/impl/file/FileWatcherNotificationHandler.java
@@ -10,6 +10,8 @@
  *******************************************************************************/
 package org.eclipse.che.api.vfs.impl.file;
 
+import org.eclipse.che.api.project.shared.dto.event.FileWatcherEventType;
+
 import java.io.File;
 
 public interface FileWatcherNotificationHandler {
