+++ wsagent/che-core-api-project/src/test/java/org/eclipse/che/api/vfs/impl/file/FileTreeWatcherTest.java
@@ -24,9 +24,9 @@ import java.util.List;
 import java.util.Set;
 
 import static com.google.common.collect.Sets.newHashSet;
-import static org.eclipse.che.api.vfs.impl.file.FileWatcherEventType.CREATED;
-import static org.eclipse.che.api.vfs.impl.file.FileWatcherEventType.DELETED;
-import static org.eclipse.che.api.vfs.impl.file.FileWatcherEventType.MODIFIED;
+import static org.eclipse.che.api.project.shared.dto.event.FileWatcherEventType.CREATED;
+import static org.eclipse.che.api.project.shared.dto.event.FileWatcherEventType.DELETED;
+import static org.eclipse.che.api.project.shared.dto.event.FileWatcherEventType.MODIFIED;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.any;
