+++ wsagent/pom.xml
@@ -25,6 +25,7 @@
     <packaging>pom</packaging>
     <name>Che Agent Parent</name>
     <modules>
+        <module>che-core-api-project-shared</module>
         <module>che-core-api-project</module>
         <module>che-core-api-git</module>
         <module>che-core-api-git-shared</module>
